pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "src"
version = "0.1.0"
description = " "
authors = [
    { name = "Bharat Aameriya", email = "bharataameriya@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "from-root==1.3.0",
    "gdown==5.2.0",
    "python-box==7.3.2",
    "python-dotenv==1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
]

[tool.setuptools]
# no packages key here

[tool.setuptools.packages.find]
where = ["."]
